+++ ecoscope/platform/tasks/results/_pydeck.py
